+++ packages/bridge/src/runtime/asyncData.ts
@@ -1,7 +1,7 @@
 import { onBeforeMount, onServerPrefetch, onUnmounted, ref, getCurrentInstance, watch } from 'vue'
 import type { Ref, WatchSource } from 'vue'
 import type { NuxtAppCompat } from '@nuxt/bridge-schema'
-import { useNuxtApp } from './app'
+import { useNuxtApp } from './nuxt'
 
 export type _Transform<Input = any, Output = any> = (input: Input) => Output
