+++ packages/bridge/src/runtime/capi.legacy.ts
@@ -2,7 +2,7 @@ import { defu } from 'defu'
 import { ComputedRef, computed, getCurrentInstance as getVM, isReactive, isRef, onBeforeMount, onServerPrefetch, reactive, ref, set, shallowRef, toRaw, toRefs, watch } from 'vue'
 import type { Route } from 'vue-router'
 import type { Nuxt2Context } from '@nuxt/bridge-schema'
-import { useNuxtApp } from './app'
+import { useNuxtApp } from './nuxt'
 import { useRouter as _useRouter, useRoute as _useRoute, useState } from './composables'
 // @ts-expect-error virtual file
 import { isFullStatic } from '#build/composition-globals.mjs'
