+++ packages/bridge/src/runtime/index.ts
@@ -1,2 +1,3 @@
 export * from './app'
+export * from './nuxt'
 export * from './composables'
